requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
pyodbc==5.1.0
simple-salesforce==1.14.0
PyYAML==6.0.1
